NOJIRA bruk rulereason / outcome for beregnet konto
