+++ src/main/java/no/nav/foreldrepenger/regler/uttak/beregnkontoer/KontoOutcome.java
@@ -1,0 +1,52 @@
+package no.nav.foreldrepenger.regler.uttak.beregnkontoer;
+
+import java.util.Map;
+
+import no.nav.fpsak.nare.evaluation.RuleReasonRef;
+
+public class KontoOutcome implements RuleReasonRef {
+    private Map<StønadskontoBeregningStønadskontotype, Integer> kontoer;
+    private Integer antallExtraBarnDager;
+    private Integer antallPrematurDager;
+
+    KontoOutcome(Map<StønadskontoBeregningStønadskontotype, Integer> kontoer) {
+        this.kontoer = kontoer;
+    }
+
+    public Map<StønadskontoBeregningStønadskontotype, Integer> getKontoer() {
+        return kontoer;
+    }
+
+    public Integer getAntallExtraBarnDager() {
+        return antallExtraBarnDager;
+    }
+
+    public Integer getAntallPrematurDager() {
+        return antallPrematurDager;
+    }
+
+    public KontoOutcome medKontoer(Map<StønadskontoBeregningStønadskontotype, Integer> kontoer) {
+        this.kontoer = kontoer;
+        return this;
+    }
+
+    public KontoOutcome medAntallExtraBarnDager(Integer antallExtraBarnDager) {
+        this.antallExtraBarnDager = antallExtraBarnDager;
+        return this;
+    }
+
+    public KontoOutcome medAntallPrematurDager(Integer antallPrematurDager) {
+        this.antallPrematurDager = antallPrematurDager;
+        return this;
+    }
+
+    @Override
+    public String getReasonTextTemplate() {
+        return "konto";
+    }
+
+    @Override
+    public String getReasonCode() {
+        return "konto";
+    }
+}
